refactor: add title property to Text model

+++ backend/src/services/text.service.ts
@@ -1,8 +1,12 @@
 import { Text } from "../models";
 import { ApiError } from "../utils";
 
-export const createText = async (content: string, creatorId: number) => {
-  return await Text.create({ content, creatorId });
+export const createText = async (
+  title: string,
+  content: string,
+  creatorId: number
+) => {
+  return await Text.create({ title, content, creatorId });
 };
 
 export const getAllTexts = async () => {
@@ -13,15 +17,30 @@ export const getTextById = async (id: number): Promise<Text | null> => {
   return await Text.findByPk(id);
 };
 
-export const updateText = async (id: number, newContent: string) => {
+export const updateText = async (
+  id: number,
+  newTitle: string,
+  newContent: string,
+  creatorId: number
+) => {
   const text = await Text.findByPk(id);
   if (!text) throw new ApiError("Text not found", 404);
 
-  text.set({ content: newContent });
+  if (text.creatorId !== creatorId) {
+    throw new ApiError("You are not authorized to update this text", 403);
+  }
+
+  text.set({ title: newTitle, content: newContent });
   return await text.save();
 };
 
-export const deleteText = async (id: number) => {
+export const deleteText = async (id: number, creatorId: number) => {
+  const text = await Text.findByPk(id);
+  if (!text) throw new ApiError("Text not found", 404);
+
+  if (text.creatorId !== creatorId) {
+    throw new ApiError("You are not authorized to delete this text", 403);
+  }
   await Text.destroy({ where: { id } });
 };
